LSOP-129 ShotOnGoal for the team (just a tiny difference to player)
git-svn-id: http://isdb1.offis.uni-oldenburg.de/repos/odysseus/trunk@20722 aa0bda67-78ed-0310-a382-c5338b0f89ba

Former-commit-id: cc338293fd454a483f37b772109864c7e05b1d83

+++ lsop/de.uniol.inf.is.odysseus.sports.sportsql/src/de/uniol/inf/is/odysseus/sports/sportsql/parser/buildhelper/OperatorBuildHelper.java
@@ -325,7 +325,7 @@ public class OperatorBuildHelper {
 	 *            source with team_id
 	 * @return SelectAO which filters for the given teamId
 	 */
-	public static SelectAO createTeamSelectAO(int teamId,
+	public static SelectAO createTeamSelectAO(long teamId,
 			ILogicalOperator source) {
 		SelectAO teamSelectAO = new SelectAO();
